simplify file names
